bibliomanager/requirements.txt
celery==5.3.6
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
redis==5.0.1
mysqlclient==2.2.0
django-cors-headers==3.8.0
djangorestframework-simplejwt==5.2.2
django-celery-beat==2.5.0